refactor(client): author card
