+++ libs/blog/articles/feature-article/src/lib/article-details/article-details.component.html
@@ -35,7 +35,7 @@
     </div>
   </section>
   <aside class="order-3 col-span-12 lg:col-span-4">
-    <al-author-info
+    <al-author-card
       [author]="articleDetails().author"
       [clampText]="true"
       [linkable]="true"
