+++ libs/blog/authors/feature-author/src/lib/feature-author/feature-author.component.html
@@ -7,9 +7,9 @@
     authorDetailsStore.authorDetails();
   as author
 ) {
-  <al-author-info [author]="author" />
+  <al-author-card [author]="author" />
 } @else {
-  <al-author-info-skeleton />
+  <al-author-card-skeleton />
 }
 
 <h2 class="mb-6 mt-10 text-2xl font-bold">Posted by {{ authorName() }}</h2>
@@ -26,9 +26,7 @@
 </div>
 
 @if (articleStore.total() > pageSize) {
-  <footer
-    class="bg-al-card mt-16 rounded-lg border px-6 py-5"
-  >
+  <footer class="bg-al-card mt-16 rounded-lg border px-6 py-5">
     <al-pagination
       alQueryPagination
       [pageSize]="pagination().take"
