+++ libs/blog/authors/ui-author-card/.eslintrc.json
@@ -24,7 +24,8 @@
             "prefix": "al",
             "style": "kebab-case"
           }
-        ]
+        ],
+        "@angular-eslint/no-host-metadata-property": "off"
       }
     },
     {
