--- libs/blog/authors/ui-author-card/src/lib/author-card.component.html
@@ -1,25 +1,0 @@
-<al-card alGradientCard>
-  <div alCardContent class="grid grid-cols-1 gap-6 md:grid-cols-[250px_auto]">
-    <al-card alGradientCard class="h-fit">
-      <div alCardContent class="flex flex-col items-center gap-2">
-        <al-avatar [imageSrc]="author().avatarUrl" size="md" />
-        <span class="whitespace-nowrap text-base font-medium">
-          {{ author().name }}
-        </span>
-        <span class="text-sm font-normal">{{ author().position }}</span>
-        <a
-          class="text-al-blue text-sm font-normal"
-          [routerLink]="['/', 'author', author().slug]"
-        >
-          Read the articles
-        </a>
-        <div class="mt-4 block text-base font-normal md:hidden">
-          {{ author().description }}
-        </div>
-      </div>
-    </al-card>
-    <div class="hidden items-center text-base font-normal md:flex">
-      {{ author().description }}
-    </div>
-  </div>
-</al-card>
