+++ libs/blog/authors/ui-author-card/src/lib/author-card/author-card-template.component.ts
@@ -7,7 +7,7 @@ import {
 } from '@angular-love/blog/shared/ui-card';
 
 @Component({
-  selector: 'al-author-info-template',
+  selector: 'al-author-card-template',
   standalone: true,
   imports: [CardComponent, CardContentDirective, GradientCardDirective],
   host: {
@@ -17,10 +17,10 @@ import {
     <al-card alGradientCard>
       <div alCardContent>
         <div
-          class="@3xl:flex-row @5xl:border-none flex  w-full flex-col items-center  rounded-lg border"
+          class="@3xl:flex-row @3xl:border-none flex  w-full flex-col items-center  rounded-lg border"
         >
           <div
-            class="@5xl:border @5xl:bg-al-gradient @5xl:min-w-[360px] min-w-fit rounded-lg pb-4 pt-6"
+            class="@3xl:border @3xl:!bg-al-radial-gradient @3xl:bg-al-background @3xl:min-w-[260px] min-w-fit rounded-lg pb-4 pt-6"
           >
             <div
               class="@3xl:max-w-[360px] flex w-full  flex-col items-center gap-4"
@@ -29,7 +29,7 @@ import {
             </div>
           </div>
 
-          <div class=" @3xl:pt-6 w-full flex-1 p-6 pt-0">
+          <div class="@3xl:pt-6 w-full flex-1 p-6 pt-0">
             <ng-content select="[author-info-description]"></ng-content>
           </div>
         </div>
@@ -37,4 +37,4 @@ import {
     </al-card>
   `,
 })
-export class AuthorInfoTemplateComponent {}
+export class AuthorCardTemplateComponent {}
