--- libs/blog/authors/ui-author-info/README.md
@@ -1,7 +1,0 @@
-# blog-authors-ui-author-info
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test blog-authors-ui-author-info` to execute the unit tests.
