--- libs/blog/authors/ui-author-info/project.json
@@ -1,20 +1,0 @@
-{
-  "name": "blog-authors-ui-author-info",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/blog/authors/ui-author-info/src",
-  "prefix": "al",
-  "projectType": "library",
-  "tags": ["scope:client", "type:ui"],
-  "targets": {
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/blog/authors/ui-author-info/jest.config.ts"
-      }
-    },
-    "lint": {
-      "executor": "@nx/eslint:lint"
-    }
-  }
-}
