--- libs/blog/authors/ui-author-info/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/author-info/author-info.component';
-export * from './lib/author-info/author-info-skeleton.component';
