--- libs/blog/authors/ui-author-info/src/test-setup.ts
@@ -1,9 +1,0 @@
-import 'jest-preset-angular/setup-jest';
-
-// @ts-expect-error https://thymikee.github.io/jest-preset-angular/docs/getting-started/test-environment
-globalThis.ngJest = {
-  testEnvironmentOptions: {
-    errorOnUnknownElements: true,
-    errorOnUnknownProperties: true,
-  },
-};
